Launched Alpha 0.1

+++ app/store/securityAreasStore.ts
@@ -10,26 +10,32 @@ import {
   deleteItemInDatabase as deleteItemFromDatabase,
   resetDatabase,
   deleteDatabase,
+  isDatabaseEmpty,
 } from '@/api/storage'
 
 import { getLocation, updateLocation, removeSecurityAreaFromLocation } from './locationsStore'
 
 import { SecurityArea, LayerVisibility, Database, Location, Alarm } from '@/api/types'
+import { CouchBase } from '@triniwiz/nativescript-couchbase'
 
 const state = Vue.observable({
   securityAreas: [] as SecurityArea[],
 })
 
 // Create persist security areas database
-const database: Database = createDatabase('securityAreas')
+let database: Database
 
-const initializeDatabase = (): void => {
+export const initializeDatabase = (name: string): void => {
+  database = createDatabase(name)
+  hydratingState(database)
+  // deleteDatabase(database)
+}
+
+export const hydratingState = (database: CouchBase) => {
   const securityAreas = getAllItemsFromDatabase(database)
-
-  if (securityAreas.length > 0) {
+  if (!isDatabaseEmpty(database))
     securityAreas.forEach((securityArea: SecurityArea) => addSecurityAreaToState(securityArea))
-  }
-  return
+  else console.log(`_______NO HAY ELEMENTOS EN LA BASE DE DATOS DE SECURITY AREAS DB______`)
 }
 
 const addSecurityAreaToState = async (securityArea: SecurityArea): Promise<void> => {
@@ -37,30 +43,40 @@ const addSecurityAreaToState = async (securityArea: SecurityArea): Promise<void>
   state.securityAreas.push(securityArea)
 }
 
-initializeDatabase()
-
-// deleteDatabase(database)
+// initializeDatabase()
 
 export const addNewSecurityArea = async (securityArea: SecurityArea): Promise<void> => {
-  console.log(
-    `securityAreaStores::addNewSecurityArea()::securityArea: ${JSON.stringify(securityArea)}`,
-  )
+  // console.log(
+  //   `securityAreaStores::addNewSecurityArea()::securityArea: ${JSON.stringify(securityArea)}`,
+  // )
   addSecurityAreaToState(securityArea)
-    .then(() => addSecurityAreaToLocation(securityArea.id))
+    .then(() => {
+      addSecurityAreaToLocation(securityArea.id, securityArea.owner)
+      console.log(`securityAreaStores::addNewSecurityArea()::securityAreaId: ${securityArea.id}`)
+    })
     .then(() => {
       addItemToDatabase<SecurityArea>(database, securityArea, securityArea.id)
+      const securityAreas = getAllItemsFromDatabase(database)
+      console.log(
+        `####_securityAreaStores::addNewSecurityArea()::securityAreas: ${JSON.stringify(
+          securityAreas,
+        )}`,
+      )
     })
 }
 
-const addSecurityAreaToLocation = async (id: string): Promise<void> => {
-  const location = setLocation(id)
+const addSecurityAreaToLocation = async (id: string, owner: string): Promise<void> => {
+  const location = setLocation(id, owner)
   updateLocation(location)
 }
 
-const setLocation = (id: string): Location => {
-  const location: Location = getLocation(id)
-  location.securityAreas.push(id)
-  return location
+const setLocation = (id: string, owner: string): Location => {
+  const location: Location = getLocation(owner)
+  if (!!location) {
+    location.securityAreas.push(id)
+    console.log(`securityAreaStores::setLocation()::securityArea id: ${id}`)
+    return location
+  } else console.log('the location not exist')
 }
 
 export const getSecurityArea = (id: string): SecurityArea =>
@@ -68,33 +84,44 @@ export const getSecurityArea = (id: string): SecurityArea =>
 
 export const getAllSecurityAreas = (): SecurityArea[] => state.securityAreas
 
-export const isId = (id: string): boolean => (findSecurityAreaIndex(id) >= 0 ? true : false)
+export const isId = (id: string): boolean => {
+  if (hasSecurityAreas()) findSecurityAreaIndex(id) >= 0
+  return
+}
 
 export const isSecurityAreaVisible = (id: string): LayerVisibility => {
-  const index = findSecurityAreaIndex(id)
-  const isVisible = state.securityAreas[index].layer.paint.visibility
-  return isVisible
+  if (hasSecurityAreas()) {
+    const index = findSecurityAreaIndex(id)
+    return state.securityAreas[index].layer.paint.visibility
+  }
+  return
 }
 
 // TODO: to review code
 export const updateSecurityAreaStore = (securityArea: SecurityArea): void => {
-  deleteSecurityArea(securityArea.id).then(() => {
-    addNewSecurityArea(securityArea)
-    updateItemInDatabase(database, securityArea.id, securityArea)
-  })
+  if (hasSecurityAreas()) {
+    deleteSecurityArea(securityArea.id).then(() => {
+      addNewSecurityArea(securityArea)
+      updateItemInDatabase(database, securityArea.id, securityArea)
+    })
+  }
+  return
 }
 
 export const deleteSecurityArea = async (id: string): Promise<void> => {
-  const index = findSecurityAreaIndex(id)
-  const owner = getOwner(index)
-  const searchId = getSearchId(index)
-  stopRadar(searchId)
-  removeSecurityAreafromState(index)
-    .then(() => removeSecurityAreaFromLocation(owner, id))
-    .then(() => {
-      deleteItemFromDatabase(database, id)
-      console.log('removed Security Area!!')
-    })
+  if (hasSecurityAreas()) {
+    const index = findSecurityAreaIndex(id)
+    const owner = getOwner(index)
+    const searchId = getSearchId(index)
+    stopRadar(searchId)
+    removeSecurityAreafromState(index)
+      .then(() => removeSecurityAreaFromLocation(owner, id))
+      .then(() => {
+        deleteItemFromDatabase(database, id)
+        console.log('removed Security Area!!')
+      })
+  }
+  return
 }
 
 const getOwner = (index: number): string => state.securityAreas[index].owner
@@ -104,8 +131,11 @@ const removeSecurityAreafromState = async (index: number): Promise<void> => {
 }
 
 export const resetSecurityAreasStore = (): void => {
-  deleteAllSecurityAreasFromState()
-  resetDatabase(database)
+  if (hasSecurityAreas()) {
+    deleteAllSecurityAreasFromState()
+    resetDatabase(database)
+  }
+  return
 }
 
 const deleteAllSecurityAreasFromState = (): void => {
@@ -115,37 +145,47 @@ const deleteAllSecurityAreasFromState = (): void => {
 const findSecurityAreaIndex = (id: string): number =>
   state.securityAreas.findIndex((securityArea) => securityArea.id === id)
 
+const hasSecurityAreas = (): boolean => !!getAllSecurityAreas() && getAllSecurityAreas().length > 0
+
 /******** ALARMS ********/
 
 export const getAlarm = (id: string): Alarm => {
-  const securityArea = getSecurityAreaFromAlarmId(id)
-  const alarm = securityArea.alarm
-  console.log(`securityAreasStore::getAlarm::alarm: ${JSON.stringify(alarm)}`)
-  return alarm
+  if (hasAlarms()) getSecurityAreaFromAlarmId(id).alarm
+  return
 }
 
 export const getAllAlarms = (): Alarm[] =>
   state.securityAreas.map((securityArea) => securityArea.alarm)
 
-export const getAlarmsActivated = (): Alarm[] =>
-  getAllAlarms().filter((alarm) => alarm.isActivated === true)
+export const getActivatedAlarms = (): Alarm[] => {
+  if (hasAlarms()) getAllAlarms().filter((alarm) => alarm.isActivated === true)
+  return
+}
 
-export const getSecurityAreaFromAlarmId = (alarmId: string): SecurityArea =>
-  state.securityAreas.find((securityArea) => securityArea.alarm.id === alarmId)
+export const getSecurityAreaFromAlarmId = (alarmId: string): SecurityArea => {
+  if (hasAlarms()) state.securityAreas.find((securityArea) => securityArea.alarm.id === alarmId)
+  return
+}
 
 export const setAlarmOn = (alarmId: string, searchId: number): void => {
-  console.log('securityAreasStore::setAlarmOn()')
-  const index = findSecurityAreaIndexFromAlarmId(alarmId)
-  setAlarmActivation(index, true)
-  setSearchId(index, searchId)
-  updateItemInDatabase(database, state.securityAreas[index].id, state.securityAreas[index])
+  if (hasAlarms()) {
+    console.log('securityAreasStore::setAlarmOn()')
+    const index = findSecurityAreaIndexFromAlarmId(alarmId)
+    setAlarmActivation(index, true)
+    setSearchId(index, searchId)
+    updateItemInDatabase(database, state.securityAreas[index].id, state.securityAreas[index])
+  }
+  return
 }
 
 export const setAlarmOff = async (alarmId: string): Promise<void> => {
-  const index = findSecurityAreaIndexFromAlarmId(alarmId)
-  setAlarmActivation(index, false)
-  setSearchId(index, null)
-  updateItemInDatabase(database, state.securityAreas[index].id, state.securityAreas[index])
+  if (hasAlarms()) {
+    const index = findSecurityAreaIndexFromAlarmId(alarmId)
+    setAlarmActivation(index, false)
+    setSearchId(index, null)
+    updateItemInDatabase(database, state.securityAreas[index].id, state.securityAreas[index])
+  }
+  return
 }
 
 const findSecurityAreaIndexFromAlarmId = (alarmId: string): number =>
@@ -156,9 +196,8 @@ const setSearchId = (index: number, searchId: number): void => {
 }
 
 export const getSearchIdFromAlarmId = (alarmId: string): number => {
-  const alarm = getAlarm(alarmId)
-  const searchId = alarm.searchId
-  return searchId
+  if (hasAlarms()) getAlarm(alarmId).searchId
+  return
 }
 
 const getSearchId = (index: number): number => state.securityAreas[index].alarm.searchId
@@ -166,3 +205,5 @@ const getSearchId = (index: number): number => state.securityAreas[index].alarm.
 const setAlarmActivation = (index: number, value: boolean): void => {
   state.securityAreas[index].alarm.isActivated = value
 }
+
+const hasAlarms = (): boolean => !!getAllAlarms() && getAllAlarms().length > 0
